cleanup of main function
