+++ src/utils/interactions/constants.rs
@@ -1,0 +1,3 @@
+use serenity::model::prelude::RoleId;
+
+pub const KF2_ROLE_ID: RoleId = RoleId(1035710963070013540);
